f32

+++ src/lib/Covariance.cpp
@@ -19,7 +19,7 @@
 //' @ref https://github.com/cran/DiceKriging/blob/master/src/CovFuns.c
 // Covariance function on normalized data
 
-std::function<double(const arma::fvec&, const arma::fvec&)> Covariance::Cov_gauss
+std::function<float(const arma::fvec&, const arma::fvec&)> Covariance::Cov_gauss
     = [](const arma::fvec& _dX, const arma::fvec& _theta) {
         const arma::fvec& dXnorm = _dX / _theta;
         return exp(-0.5 * arma::dot(dXnorm, dXnorm));
@@ -35,7 +35,7 @@ std::function<arma::fvec(const arma::fvec&, const arma::fvec&)> Covariance::DlnC
         return arma::conv_to<arma::fcolvec>::from(-_dX / arma::square(_theta));
       };
 
-std::function<double(const arma::fvec&, const arma::fvec&)> Covariance::Cov_exp
+std::function<float(const arma::fvec&, const arma::fvec&)> Covariance::Cov_exp
     = [](const arma::fvec& _dX, const arma::fvec& _theta) { return exp(-arma::sum(arma::abs(_dX / _theta))); };
 
 std::function<arma::fvec(const arma::fvec&, const arma::fvec&)> Covariance::DlnCovDtheta_exp
@@ -48,9 +48,9 @@ std::function<arma::fvec(const arma::fvec&, const arma::fvec&)> Covariance::DlnC
         return arma::conv_to<arma::fcolvec>::from(-arma::sign(_dX) / _theta);
       };
 
-const double SQRT_3 = std::sqrt(3.0);
+const float SQRT_3 = std::sqrt(3.0);
 
-std::function<double(const arma::fvec&, const arma::fvec&)> Covariance::Cov_matern32
+std::function<float(const arma::fvec&, const arma::fvec&)> Covariance::Cov_matern32
     = [](const arma::fvec& _dX, const arma::fvec& _theta) {
         arma::fvec d = SQRT_3 * arma::abs(_dX / _theta);
         return exp(-arma::sum(d - arma::log1p(d)));
@@ -68,9 +68,9 @@ std::function<arma::fvec(const arma::fvec&, const arma::fvec&)> Covariance::DlnC
         return arma::conv_to<arma::fcolvec>::from(-SQRT_3 * arma::sign(_dX) % d / (1 + d) / _theta);
       };
 
-const double SQRT_5 = std::sqrt(5.0);
+const float SQRT_5 = std::sqrt(5.0);
 
-std::function<double(const arma::fvec&, const arma::fvec&)> Covariance::Cov_matern52
+std::function<float(const arma::fvec&, const arma::fvec&)> Covariance::Cov_matern52
     = [](const arma::fvec& _dX, const arma::fvec& _theta) {
         arma::fvec d = SQRT_5 * arma::abs(_dX / _theta);
         return exp(-arma::sum(d - arma::log1p(d + (d % d) / 3)));
@@ -92,9 +92,9 @@ std::function<arma::fvec(const arma::fvec&, const arma::fvec&)> Covariance::DlnC
         return arma::conv_to<arma::fcolvec>::from(-SQRT_5 * arma::sign(_dX) % (a % b) / (a + d % b) / _theta);
       };
 
-const double EPSILON = 1E-13;
+const float EPSILON = 1E-13;
 
-std::function<double(const arma::fvec&, const arma::fvec&)> Covariance::Cov_whitenoise
+std::function<float(const arma::fvec&, const arma::fvec&)> Covariance::Cov_whitenoise
     = [](const arma::fvec& _dX, const arma::fvec& _theta) {
         if (arma::sum(arma::abs(_dX / _theta)) < EPSILON)
           return 1.0;
